requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
# Flask==2.1.2
asgiref==3.3.4; python_version >= '3.6'
cycler==0.10.0
django==3.2.4
kiwisolver==1.3.1; python_version >= '3.6'
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
pillow==8.2.0
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
typing-extensions==3.10.0.0; python_version < '3.8'